==> 03_analyze full simulation data_manuscript.Rmd <==
---
title: "Analyze Data from Monte Carlo Simulation"
author: "Chase Latour"
date: "`r Sys.Date()`"
output: 
  html_document:
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(knitr)
library(kableExtra)
library(showtext)
font_add("Calibri", "/Library/Fonts/calibri.ttf") 

# Enable showtext for rendering
showtext_auto()

# library(extrafont)
# font_import(paths = "/Library/Fonts", pattern = "calibri.ttf", prompt=FALSE)
# loadfonts(device = "pdf")

```

# Pull in the Data

```{r}

# CHANGE THE 6 TO 12 WHEN DONE
nscenario <- 12

# Descriptive data from each of the scenarios

for (x in 1:nscenario) {
  # Construct the file name
  file_name <- paste0("describe trial_scenario", x, ".rds")
  
  # Read the RDS file
  dataset <- readRDS(file_name)
  
  # Add the new variable 'scenario'
  dataset$scenario <- x
  
  # Construct the variable name
  var_name <- paste0("describe_scenario", x)
  
  # Assign the dataset to the dynamically named variable
  assign(var_name, dataset)
  
}

combined_describe_trial <- bind_rows(describe_scenario1, describe_scenario2, describe_scenario3,
                                     describe_scenario4, describe_scenario5, describe_scenario6,
                                     describe_scenario7, describe_scenario8, describe_scenario9,
                                     describe_scenario10, describe_scenario11, describe_scenario12)

# Descriptive outcome data from each of the scenarios

for (x in 1:nscenario) {
  
  file_name <- paste0("describe outcomes_scenario", x, ".rds")
  
  dataset <- readRDS(file_name)
  
  dataset$scenario <- x
  
  var_name <- paste0("describe_scenario", x)
  
  assign(var_name, dataset)
  
}

combined_describe_outcomes <- bind_rows(describe_scenario1, describe_scenario2, describe_scenario3,
                                     describe_scenario4, describe_scenario5, describe_scenario6,
                                     describe_scenario7, describe_scenario8, describe_scenario9,
                                     describe_scenario10, describe_scenario11, describe_scenario12)

# Analysis data for each of the scenarios

for (x in 1:nscenario) {
  # Construct the file name
  file_name <- paste0("analyses_scenario", x, ".rds")
  
  # Read the RDS file
  dataset <- readRDS(file_name)
  
  # Add the new variable 'scenario'
  dataset$scenario <- x
  
  # Construct the variable name
  var_name <- paste0("analyses_scenario", x)
  
  # Assign the dataset to the dynamically named variable
  assign(var_name, dataset)
}

combined_analyses <- bind_rows(analyses_scenario1, analyses_scenario2, analyses_scenario3,
                               analyses_scenario4, analyses_scenario5, analyses_scenario6,
                               analyses_scenario7, analyses_scenario8, analyses_scenario9,
                               analyses_scenario10, analyses_scenario11, analyses_scenario12)

```

# Descriptives of Trial Population

```{r}

n_outcomes <- combined_describe_trial %>% 
  group_by(scenario, trt) %>% 
  summarize(n_pregnancies = median(n_people),
            n_censored = median(n_censor),
            n_fetal_deaths = median(n_fd_no_censor),
            n_live_births = median(n_lb_no_censor),
            n_preeclampsia = median(n_preeclampsia_no_cen),
            n_sga = median(n_sga_no_cen),
            .groups = 'drop')

totals_by_scenario <- n_outcomes %>% 
  group_by(scenario) %>% 
  summarize(total_pregnancies = sum(n_pregnancies),
            total_censored = sum(n_censored),
            total_fetal_deaths = sum(n_fetal_deaths),
            total_live_births = sum(n_live_births),
            total_preeclampsia = sum(n_preeclampsia),
            total_sga = sum(n_sga),
            .groups = 'drop')


descriptives <- left_join(n_outcomes, totals_by_scenario, by = ("scenario" = "scenario"))

descriptives2 <- descriptives %>% 
  rowwise() %>% 
  mutate(#perc_pregnancies = round(100*n_pregnancies / total_pregnancies, 1),
         perc_censored = round(100*n_censored / n_pregnancies, 1),
         perc_fetal_deaths = round(100*n_fetal_deaths / n_pregnancies, 1),
         perc_live_births = round(100*n_live_births / n_pregnancies, 1),
         perc_preeclampsia = round(100*n_preeclampsia / n_pregnancies, 1),
         perc_sga = round(100*n_sga / n_pregnancies, 1)) %>% 
  ungroup() %>% 
  mutate(pregnancies = n_pregnancies, #paste0(n_pregnancies, " (", perc_pregnancies, "%)"),
         censored = paste0(n_censored, " (", perc_censored, "%)"),
         fetal_deaths = paste0(n_fetal_deaths, " (", perc_fetal_deaths, "%)"),
         live_births = paste0(n_live_births, " (", perc_live_births, "%)"),
         preeclampsia = paste0(n_preeclampsia, " (", perc_preeclampsia, "%)"),
         sga = paste0(n_sga, " (", perc_sga, "%)")) %>% 
  select(scenario, trt, pregnancies, censored, fetal_deaths, live_births, preeclampsia, sga) %>% 
  pivot_wider(id_cols = scenario, 
              names_from = trt,
              values_from = c(pregnancies, censored, fetal_deaths, live_births, preeclampsia, sga))
# library(openxlsx)
# write.xlsx(descriptives2, file = "descriptive_results.xlsx")
  
descriptives2 %>% 
  knitr::kable(digits = 3, format.args = list(big.mark = ",", scientific = FALSE),
               col.names = c()) %>% # Trying to add commas to larger numbers
  kable_classic() %>% 
  add_header_above(c("Scenario" = 1, "Untreated" = 1, "Treated" = 1, 
                     "Untreated" = 1, "Treated" = 1,
                     "Untreated" = 1, "Treated" = 1, 
                     "Untreated" = 1, "Treated" = 1,
                     "Untreated" = 1, "Treated" = 1,
                     "Untreated" = 1, "Treated" = 1)) %>% 
  add_header_above(c(" " = 1, "Number of Pregnancies" = 2, 
                     "Number of Censored Pregnancies" = 2,
                     "Number of Fetal Deaths" = 2, 
                     "Number of Live Births" = 2,
                     "Number of Preeclampsia" = 2,
                     "Number of SGA" = 2)) 

```

# Timing Descriptives of Trial Population

```{r}

combined_describe_outcomes %>%
  filter(strata == "Total") %>%
  group_by(scenario, trt, preg_outcome_final) %>%
  summarize(min_median = min(median), median_median = median(median), max_median = max(median), .groups = "drop") 
```

# Results


```{r}

# Establish the true estimates as the average of the potential outcomes

truth <- combined_analyses %>% 
  group_by(scenario) %>% 
  summarize(composite_truth_r1 = mean(`potential_ risk_composite_trt`),
            composite_truth_r0 = mean(`potential_ risk_composite_untrt`),
            composite_truth_rd = mean(`potential_ rd_composite`),
            composite_truth_rr = exp(mean(log(`potential_ rr_composite`))),
            sga_truth_r1 = mean(`potential_ risk_sga_trt`),
            sga_truth_r0 = mean(`potential_ risk_sga_untrt`),
            sga_truth_rd = mean(`potential_ rd_sga`),
            sga_truth_rr = exp(mean(log(`potential_ rr_sga`))))

combined_analyses2 <- left_join(combined_analyses, truth, by = "scenario")


calculate_simulation_values <- function(dataset){
  
  hold <- dataset %>% 
    group_by(scenario) %>% 
    summarize(
      
      # True Values
      truth_r0 = mean(r0_truth)*100,
      truth_r1 = mean(r1_truth)*100,
      truth_rd = mean(rd_truth)*100,
      truth_rr = exp(mean(log(rr_truth))),
      
      # Kaplan-Meier Estimator
      ## Absolute risk
      km_r0 = mean(kmr0)*100,
      km_r1 = mean(kmr1)*100,
      ## Risk Difference
      km_rd = mean(kmrd)*100, # Estimate
      km_rd_se = sqrt((1 / (n()*(n()-1)))*
                        sum((kmrd - km_rd)^2)),
      km_rd_bias = km_rd - truth_rd, #mean(kmrd) - truth_rd,
      km_rd_ese = sqrt(var(kmrd)),
      km_rd_rmse = sqrt(mean(((kmrd*100) - truth_rd)^2)),
      ## Risk Ratio
      km_rr = exp(mean(log(kmrr))),
      km_lnrr_se = sqrt((1 / (n()*(n()-1)))*
                          sum((log(kmrr) - log(km_rr))^2)),
      km_lnrr_bias = mean(log(kmrr)) - log(truth_rr),
      km_lnrr_ese = sqrt(var(log(kmrr))),
      km_lnrr_rmse = sqrt(mean((log(kmrr) - log(truth_rr))^2)),
      
      
      # Aalen-Johansen Estimator
      ## Absolute risk
      aj_r0 = mean(ajr0)*100,
      aj_r1 = mean(ajr1)*100,
      ## Risk Difference
      aj_rd = mean(ajrd)*100,
      aj_rd_se = sqrt((1 / (n()*(n()-1)))*
                        sum((ajrd - aj_rd)^2)),
      aj_rd_bias =aj_rd - truth_rd,
      aj_rd_ese = sqrt(var(ajrd)),
      aj_rd_rmse = sqrt(mean(((100*ajrd) - truth_rd)^2)),
      ## Risk Ratio
      aj_rr = exp(mean(log(ajrr))),
      aj_lnrr_se = sqrt((1 / (n()*(n()-1)))*
                          sum((log(ajrr) - log(aj_rr))^2)),
      aj_lnrr_bias = mean(log(ajrr)) - log(truth_rr),
      aj_lnrr_ese = sqrt(var(log(ajrr))),
      aj_lnrr_rmse = sqrt(mean((log(ajrr) - log(truth_rr))^2)),
      
      .groups = "keep"
      
    )
  
  return(hold)
  
}
  

# Get the estimates for the composite outcome

composite_outcome <- combined_analyses2 %>% 
  rename(kmr0 = `km_composite_ r_0`,
         kmr1 = `km_composite_ r_1`,
         kmrd = `km_composite_ rd`,
         kmrr = `km_composite_ rr`,
         ajr0 = `composite_multi_gw r_0`,
         ajr1 = `composite_multi_gw r_1`,
         ajrd = `composite_multi_gw rd`,
         ajrr = `composite_multi_gw rr`,
         r1_truth = composite_truth_r1,
         r0_truth = composite_truth_r0,
         rd_truth = composite_truth_rd,
         rr_truth = composite_truth_rr) %>% 
  calculate_simulation_values() %>% 
  mutate(Outcome = "Composite")
  
# SGA - FD only
sga_fd_only <- combined_analyses2 %>% 
  rename(kmr0 = `km_sga_fd_ r_0`,
         kmr1 = `km_sga_fd_ r_1`,
         kmrd = `km_sga_fd_ rd`,
         kmrr = `km_sga_fd_ rr`,
         ajr0 = `aj_sga_fd_ r_0`,
         ajr1 = `aj_sga_fd_ r_1`,
         ajrd = `aj_sga_fd_ rd`,
         ajrr = `aj_sga_fd_ rr`,
         r1_truth = sga_truth_r1,
         r0_truth = sga_truth_r0,
         rd_truth = sga_truth_rd,
         rr_truth = sga_truth_rr) %>% 
  calculate_simulation_values() %>% 
  mutate(Outcome  = "SGA: 1 Competing Event")

# SGA - All competing events
sga_all <- combined_analyses2 %>% 
  rename(kmr0 = `km_sga_all_ r_0`,
         kmr1 = `km_sga_all_ r_1`,
         kmrd = `km_sga_all_ rd`,
         kmrr = `km_sga_all_ rr`,
         ajr0 = `aj_sga_all_ r_0`,
         ajr1 = `aj_sga_all_ r_1`,
         ajrd = `aj_sga_all_ rd`,
         ajrr = `aj_sga_all_ rr`,
         r1_truth = sga_truth_r1,
         r0_truth = sga_truth_r0,
         rd_truth = sga_truth_rd,
         rr_truth = sga_truth_rr) %>% 
  calculate_simulation_values() %>% 
  mutate(Outcome  = "SGA: 2 Competing Events")

# # Print in Excel file
# analysis <- rbind(composite_outcome, sga_fd_only, sga_all) %>%
#   select(-c(Outcome, contains("_se"))) %>%
#   mutate(across(where(is.numeric), ~ round(.x, 2)))
# library(openxlsx)
# write.xlsx(analysis, file = "analysis results.xlsx")

## Then stack and print
rbind(composite_outcome, sga_fd_only, sga_all) %>% 
  select(-c(Outcome, contains("_se"))) %>% 
  mutate(across(where(is.numeric), ~ round(.x, 2))) %>%
  kbl() %>% 
  kable_classic() %>% 
  pack_rows("Composite Outcome", 1, nscenario) %>% 
  pack_rows("SGA with Fetal Death as Only Competing Events", nscenario+1, 2*nscenario) %>% 
  pack_rows("SGA with 2 Competing Events", 2*nscenario+1, 3*nscenario) %>% 
  add_header_above(c(" " = 1, "Risk Estimates" = 2, "Risk Difference" = 1,
                     "Risk Ratio" = 1, "Risk Estimates" = 2, "Risk Difference" = 4, "Risk Ratio" = 4,
                     "Risk Estimates" = 2, "Risk Difference" = 4, "Risk Ratio" = 4)) %>% 
  add_header_above(c(" " =  1, "Truth" = 4, "Kaplan-Meier Estimator" = 10,
                   "Aalen-Johansen Estimator" = 10))
  

```

# Figure: Composite Outcome

## Monte Carlo Standard Error

```{r}

calc_mcse <- function(dsn, r0, r1, rd, rr, nsim=2000){
  
  mcse <- dsn %>% 
    group_by(scenario) %>% 
    summarize(
      r0_mcse = sqrt( (1/(nsim * (nsim-1))) * sum((({{r0}}*100) - (100*mean({{r0}})))^2)),
      r1_mcse = sqrt( (1/(nsim * (nsim-1))) * sum((({{r1}}*100) - (100*mean({{r1}})))^2)),
      rd_mcse = sqrt( (1/(nsim * (nsim-1))) * sum((({{rd}}*100) - (100*mean({{rd}})))^2)),
      rr_mcse = sqrt( (1/(nsim * (nsim-1))) * sum((log({{rr}}) - mean(log({{rr}})))^2)),
      .groups = 'drop'
    )
  
  return(mcse)
  
}

### Composite outcome

# KM
composite_km <- calc_mcse(dsn=combined_analyses, r0=`km_composite_ r_0`, r1=`km_composite_ r_1`,
                          rd = `km_composite_ rd`, rr=`km_composite_ rr`) %>% 
  rename_with(~ paste0("km_", .), c(r0_mcse, r1_mcse, rd_mcse, rr_mcse))
# AJ
composite_aj <- calc_mcse(dsn=combined_analyses, r0=`composite_multi_gw r_0`,r1=`composite_multi_gw r_1`,
                         rd=`composite_multi_gw rd`, rr=`composite_multi_gw rr`) %>% 
  mutate(Outcome = "Composite") %>% 
  rename_with(~ paste0("aj_", .), c(r0_mcse, r1_mcse, rd_mcse, rr_mcse))

# Left join these together
composite_mcse <- left_join(composite_km, composite_aj, by="scenario")


### SGA outcome: Censoring healthy live birth
# KM
sga_fd_km <- calc_mcse(dsn=combined_analyses, r0=`km_sga_all_ r_0`,r1=`km_sga_all_ r_1`,
                    rd=`km_sga_all_ rd`, rr=`km_sga_all_ rr`) %>% 
  rename_with(~ paste0("km_", .), c(r0_mcse, r1_mcse, rd_mcse, rr_mcse))
# AJ, censoring non-SGA live birth
sga_fd_aj <- calc_mcse(dsn=combined_analyses, r0=`aj_sga_fd_ r_0`, r1=`aj_sga_fd_ r_1`,
                       rd=`aj_sga_fd_ rd`, rr=`aj_sga_fd_ rr`) %>% 
  mutate(Outcome = "SGA: 1 Competing Event") %>% 
  rename_with(~ paste0("aj_", .), c(r0_mcse, r1_mcse, rd_mcse, rr_mcse))

# Left join these together
sga_fd_mcse <- left_join(sga_fd_km, sga_fd_aj, by = "scenario")

### SGA Outcome: Treating healthy live birth as competing event

sga_all_km <- sga_fd_km 
# AJ, only administrative censoring
sga_all_aj <- calc_mcse(dsn=combined_analyses, r0=`aj_sga_all_ r_0`, r1=`aj_sga_all_ r_1`,
                    rd=`aj_sga_all_ rd`, rr=`aj_sga_all_ rr`) %>% 
  mutate(Outcome = "SGA: 2 Competing Events") %>% 
  rename_with(~ paste0("aj_", .), c(r0_mcse, r1_mcse, rd_mcse, rr_mcse))

# Left join these together

sga_all_mcse <- left_join(sga_all_km, sga_all_aj, by = "scenario")


# Finally, stack these all together
mcse <- rbind(composite_mcse, sga_fd_mcse) %>% 
  rbind(sga_all_mcse)
  

```

## Figure

```{r}

z <- qnorm(0.975)

figure_data <- rbind(composite_outcome, sga_fd_only, sga_all) %>% 
  select(-starts_with("truth_")) %>% 
  select(-c(contains("_rd_se"), contains("_rd_bias"), contains("rd_ese"), contains("rd_rmse"),
            contains("rr_se"), contains("rr_bias"), contains("rr_ese"), contains("rr_rmse"))) %>% 
  # Left join the MCSEs
  left_join(mcse, by = c("scenario", "Outcome")) %>% 
  # Now calculate confidence intervals
  mutate(
    km_r0_lcl = km_r0 - z*km_r0_mcse,
    km_r0_ucl = km_r0 + z*km_r0_mcse,
    km_r1_lcl = km_r1 - z*km_r1_mcse,
    km_r1_ucl = km_r1 + z*km_r1_mcse,
    km_rd_lcl = km_rd - z*km_rd_mcse,
    km_rd_ucl = km_rd + z*km_rd_mcse,
    km_rr_lcl = exp(log(km_rr) - z*km_rr_mcse),
    km_rr_ucl = exp(log(km_rr) + z*km_rr_mcse),
    
    aj_r0_lcl = aj_r0 - z*aj_r0_mcse,
    aj_r0_ucl = aj_r0 + z*aj_r0_mcse,
    aj_r1_lcl = aj_r1 - z*aj_r1_mcse,
    aj_r1_ucl = aj_r1 + z*aj_r1_mcse,
    aj_rd_lcl = aj_rd - z*aj_rd_mcse,
    aj_rd_ucl = aj_rd + z*aj_rd_mcse,
    aj_rr_lcl = exp(log(aj_rr) - z*aj_rr_mcse),
    aj_rr_ucl = exp(log(aj_rr) + z*aj_rr_mcse)
  ) %>% 
  select(-c(contains("_mcse")))

figure_efficacy <- figure_data %>% 
  filter(Outcome == "Composite") %>% 
  filter(scenario <= 6) # The first and second 6 scenarios are the exact same for the composite outcome

figure_efficacy2 <- figure_efficacy %>% 
  pivot_longer(cols = -c(scenario, Outcome),
               names_to = "name",
               values_to = "value") %>% 
  mutate(
    Estimator = ifelse(grepl("km", name),
                       "KM",
                       "AJ"),
    parameter = case_when(grepl("r0_lcl", name) ~ "risk_non_initiators_lcl",
                          grepl("r0_ucl", name) ~ "risk_non_initiators_ucl",
                          grepl("_r0", name) ~ "risk_non_initiators",
                          grepl("r1_lcl", name) ~ "risk_initiators_lcl",
                          grepl("r1_ucl", name) ~ "risk_initiators_ucl",
                          grepl("_r1", name) ~ "risk_initiators",
                          grepl("rd_lcl", name) ~ "RD_LCL",
                          grepl("rd_ucl", name) ~ "RD_UCL",
                          grepl("rd", name) ~ "RD",
                          grepl("rr_lcl", name) ~ "RR_LCL",
                          grepl("rr_ucl", name) ~ "RR_UCL",
                          grepl("rr", name) ~"RR"
    )
  ) %>% 
  pivot_wider(
    id_cols = c(scenario, Outcome, Estimator),
    names_from = parameter,
    values_from = value
  ) %>% 
  left_join(truth, by = "scenario") %>% 
  mutate(
    composite_truth_r1 = 100*composite_truth_r1,
    composite_truth_r0 = 100*composite_truth_r0,
    composite_truth_rd = 100*composite_truth_rd,
    sga_truth_r1 = 100*sga_truth_r1,
    sga_truth_r0 = 100*sga_truth_r0,
    sga_truth_rd = 100*sga_truth_rd
  ) %>% 
  mutate(rd_absolute_bias = round(abs(composite_truth_rd - RD),1),
         rr_absolute_bias = round(abs(log(composite_truth_rr) - log(RR)), 2)) 

# Create a data frame with unique composite_truth_rd values for each facet
unique_rd_values <- figure_efficacy2 %>% 
  group_by(scenario) %>% 
  summarise(composite_truth_rd = unique(composite_truth_rd),
            composite_truth_r0 = unique(composite_truth_r0),
            composite_truth_r1 = unique(composite_truth_r1))

unique_rr_values <- figure_efficacy2 %>% 
  group_by(scenario) %>% 
  summarise(composite_truth_rr = unique(composite_truth_rr))


#####
## Print out the risks

pdf("Primary efficacy outcome - risks.pdf", height = 6, width = 4.4)

# BAR with truth as point estimates
ggplot(figure_efficacy2) +
  # Incidence bars for Drug A and Comparator
  geom_bar(aes(y = Estimator, x = risk_initiators, fill = "Initiators"), stat = "identity", 
           position = position_nudge(y = 0.2), width = 0.4) +
  geom_bar(aes(y = Estimator, x = risk_non_initiators, fill = "Non-Initiators"), stat = "identity", 
           position = position_nudge(y = -0.2), width = 0.4) +
  # Confidence Intervals for Initiators
  geom_errorbar(aes(y = Estimator, xmin = risk_initiators_lcl, xmax = risk_initiators_ucl), 
                width = 0.2, position = position_nudge(y = 0.2)) +
  # Confidence Intervals for Non-Initiators
  geom_errorbar(aes(y = Estimator, xmin = risk_non_initiators_lcl, xmax = risk_non_initiators_ucl), 
                width = 0.2, position = position_nudge(y = -0.2)) +
  facet_wrap(scenario ~ ., ncol = 1, strip.position = "left") +
  geom_point(data = figure_efficacy2, aes(y = Estimator, x = composite_truth_r1),
             position = position_nudge(y = 0.2), size = 3, shape = 23, fill = "white", color = "black") +
  geom_point(data = figure_efficacy2, aes(y = Estimator, x = composite_truth_r0),
             position = position_nudge(y = -0.2), size = 3, shape = 23, fill = "white", color = "black") +
  labs(y = "Scenario", 
       x = "Absolute Risk (%)",
       fill = "Treatment Group", 
       ) +
  theme_minimal() +
  theme(legend.position = "bottom",
        axis.text.y = element_text(size = 14, family = "Calibri", color="black"),
        axis.text.x = element_text(size=14, family = "Calibri", color="black"),
        strip.text = element_text(size = 14, family = "Calibri", color="black"),
        legend.text = element_text(size = 12, family = "Calibri", color="black"),
        axis.title.x = element_text(family = "Calibri", size=14, color="black"),
        axis.title.y = element_text(family = "Calibri", size=14, color="black"),
        plot.title = element_text(family = "Calibri", size=14, color="black"),
        legend.title = element_text(family = "Calibri", size=12, color="black"),
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", linewidth = 0.3),
        panel.grid.minor.x = element_line(color = "grey70", linewidth = 0.3),
        ) +
  scale_fill_manual(values = c("Initiators" = "#E69F00", "Non-Initiators" = "#0072B2")) +
  theme(strip.placement = "outside") 

dev.off()


pdf("Primary efficacy outcome - RDs.pdf", height = 6, width = 3.1)

ggplot(figure_efficacy2) +
  geom_point(aes(y = Estimator, x = RD)) +
  # Horizontal error bars for confidence intervals
  geom_errorbarh(aes(y = Estimator, xmin = RD_LCL, xmax = RD_UCL), height = 0.1, color = "black") +
  geom_vline(data = unique_rd_values, aes(xintercept = composite_truth_rd), linetype = "dashed", color = "red") +
   geom_label(aes(y = Estimator, x = 1.2, label = scales::number(rd_absolute_bias, accuracy = 0.1)), fill = "white", label.size = NA, family = "Calibri", size=3.7, color="black") +
  facet_wrap(scenario ~ ., ncol = 1, strip.position = "left") +
  labs(y = "Scenario", 
       x = "Risk Difference (per 100)",
       fill = "Treatment Group", 
       ) +
  scale_x_continuous(limits = c(-10, 1.25),
                     breaks = seq(-10, 0, by  = 5)) +
  theme_minimal() +
  theme(legend.position = "bottom",
        axis.text.y = element_text(size = 14, family = "Calibri", color="black"),
        axis.text.x = element_text(size = 14, family = "Calibri", color="black"),
        strip.text = element_text(size = 14, family = "Calibri", color="black"),
        # legend.text = element_text(size = 10, family = "Calibri"),
        axis.title.x = element_text(family = "Calibri", size=14, color="black"),
        axis.title.y = element_text(family = "Calibri", size=14, color="black"),
        plot.title = element_text(family = "Calibri", size=14, color="black"),
        axis.title.x.bottom = element_text(hjust = 0.4, color="black"),
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", linewidth = 0.3),
        panel.grid.minor.x = element_line(color = "grey70", linewidth = 0.3),
        ) +
  theme(strip.placement = "outside") 

dev.off()



pdf("Primary efficacy outcome - RRs.pdf", height = 6, width = 3.1)

ggplot(figure_efficacy2) +
  geom_point(aes(y = Estimator, x = RR)) +
  geom_errorbarh(aes(y = Estimator, xmin = RR_LCL, xmax = RR_UCL), height = 0.1, color = "black") +
  geom_label(aes(y = Estimator, x = 1.06, label = scales::number(rr_absolute_bias, accuracy = 0.01)), fill = "white", label.size = NA, family="Calibri", size=3.7, color="black") +
  geom_vline(data = unique_rr_values, aes(xintercept = composite_truth_rr), linetype = "dashed", color = "red") +
  facet_wrap(scenario ~ ., ncol = 1, strip.position = "left") +
  labs(y = "Scenario", 
       x = "Risk Ratio (log-scale)",
       fill = "Treatment Group", 
       ) +
  scale_x_continuous(
    trans='log10',
    limits = c(0.7, 1.09),
    breaks = seq(0.7, 1, by = 0.1),
    # breaks = scales::trans_breaks("log10", function(x) 10^x),
    labels = scales::label_number(accuracy = 0.1)
  ) +
  theme_minimal() +
  theme(legend.position = "bottom",
        axis.text.y = element_text(size = 14, family = "Calibri", color="black"),
        axis.text.x = element_text(size = 14, family = "Calibri", color="black"),
        strip.text = element_text(size = 14, family = "Calibri", color="black"),
        legend.text = element_text(size = 14, family = "Calibri", color="black"),
        axis.title.x = element_text(size=14, family = "Calibri", color="black"),
        axis.title.y = element_text(size=14, family = "Calibri", color="black"),
        plot.title = element_text(size=14, family = "Calibri", color="black"),
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", linewidth = 0.3),
        panel.grid.minor.x = element_line(color = "grey70", linewidth = 0.3),
        ) +
  theme(strip.placement = "outside") 

dev.off()









```

# Figure: SGA Outcome

## Scenarios 1-6

```{r}


z <- qnorm(0.975)


## Now, we only want 1 set of KM estimates for each scenario and then each of teh AJ estimates. 
## Split this up

km_safety <- figure_data %>% 
  filter(Outcome != "Composite") %>%  
  # filter(scenario <= 3) %>% 
  select(scenario, contains("km")) %>% 
  unique() %>% 
  mutate(Estimator = "KM") %>% 
  rename(risk_non_initiators = km_r0,
         risk_non_initiators_ucl = km_r0_ucl,
         risk_non_initiators_lcl = km_r0_lcl,
         risk_initiators = km_r1,
         risk_initiators_ucl = km_r1_ucl,
         risk_initiators_lcl = km_r1_lcl,
         RD = km_rd,
         RR = km_rr,
         RD_LCL = km_rd_lcl,
         RD_UCL = km_rd_ucl,
         RR_LCL = km_rr_lcl,
         RR_UCL = km_rr_ucl)

aj_safety <- figure_data %>% 
  filter(Outcome != "Composite") %>% 
  # filter(scenario <= 3) %>% 
  select(scenario, Outcome, contains("aj")) %>% 
  mutate(Estimator = ifelse(grepl("1", Outcome),
                            "AJ: 1 Competing Event",
                            "AJ: 2 Competing Events")) %>% 
  select(-Outcome) %>% 
  rename(risk_non_initiators = aj_r0,
         risk_non_initiators_ucl = aj_r0_ucl,
         risk_non_initiators_lcl = aj_r0_lcl,
         risk_initiators = aj_r1,
         risk_initiators_ucl = aj_r1_ucl,
         risk_initiators_lcl = aj_r1_ucl,
         RD = aj_rd,
         RR = aj_rr,
         RD_LCL = aj_rd_lcl,
         RD_UCL = aj_rd_ucl,
         RR_LCL = aj_rr_lcl,
         RR_UCL = aj_rr_ucl)

figure_safety <- bind_rows(km_safety, aj_safety) %>% 
  left_join(truth, by = "scenario") %>% 
  filter(scenario <= 6) %>% 
  mutate(
    composite_truth_r1 = 100*composite_truth_r1,
    composite_truth_r0 = 100*composite_truth_r0,
    composite_truth_rd = 100*composite_truth_rd,
    sga_truth_r1 = 100*sga_truth_r1,
    sga_truth_r0 = 100*sga_truth_r0,
    sga_truth_rd = 100*sga_truth_rd
  ) %>% 
  mutate(rd_absolute_bias = round(abs(sga_truth_rd - RD),1),
         rr_absolute_bias = round(abs(log(sga_truth_rr) - log(RR)), 2)) %>% 
  mutate(Estimator = factor(Estimator,
                            levels = c("AJ: 2 Competing Events",
                                       "AJ: 1 Competing Event",
                                       "KM")))

# Create a data frame with unique composite_truth_rd values for each facet
unique_rd_values <- figure_safety %>% 
  group_by(scenario) %>% 
  summarise(sga_truth_rd = unique(sga_truth_rd),
            sga_truth_r0 = unique(sga_truth_r0),
            sga_truth_r1 = unique(sga_truth_r1),
            sga_truth_rr = unique(sga_truth_rr))




#####
## Print out the risks

pdf("Primary safety outcome - risks - 1-6.pdf", height = 6, width = 4.25)

ggplot(figure_safety) +
  # Incidence bars for Drug A and Comparator
  geom_bar(aes(y = Estimator, x = risk_initiators, fill = "Initiators"), stat = "identity", 
           position = position_nudge(y = 0.2), width = 0.4) +
  geom_bar(aes(y = Estimator, x = risk_non_initiators, fill = "Non-Initiators"), stat = "identity", 
           position = position_nudge(y = -0.2), width = 0.4) +
  # Confidence Intervals for Initiators
  geom_errorbar(aes(y = Estimator, xmin = risk_initiators_lcl, xmax = risk_initiators_ucl), 
                width = 0.2, position = position_nudge(y = 0.2)) +
  # Confidence Intervals for Non-Initiators
  geom_errorbar(aes(y = Estimator, xmin = risk_non_initiators_lcl, xmax = risk_non_initiators_ucl), 
                width = 0.2, position = position_nudge(y = -0.2)) +
  facet_wrap(scenario ~ ., ncol = 1, strip.position = "left") +
  geom_point(data = figure_safety, aes(y = Estimator, x = sga_truth_r1),
             position = position_nudge(y = 0.2), size = 2, shape = 23, fill = "white", color = "black") +
  geom_point(data = figure_safety, aes(y = Estimator, x = sga_truth_r0),
             position = position_nudge(y = -0.2), size = 2, shape = 23, fill = "white", color = "black") +
  labs(y = "Scenario", 
       x = "Absolute Risk (%)",
       fill = "Treatment Group", 
       ) +
  theme_minimal() +
  theme(legend.position = "bottom",
        # axis.title.x.top = element_text(hjust = 0.97),
        axis.text.y = element_text(size = 14, family = "Calibri", color="black"),
        axis.text.x = element_text(size = 14, family = "Calibri", color="black"),
        strip.text = element_text(size = 14, family = "Calibri", color="black"),
        legend.text = element_text(size = 14, family = "Calibri", color="black"),
        axis.title.x = element_text(size=14, family = "Calibri", color="black"),
        axis.title.y = element_text(size=14, family = "Calibri", color="black"),
        plot.title = element_text(size=14, family = "Calibri", color="black"),
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", linewidth = 0.3),
        panel.grid.minor.x = element_line(color = "grey70", linewidth = 0.3),
        ) +
  scale_fill_manual(values = c("Initiators" = "#E69F00", "Non-Initiators" = "#0072B2")) +
  theme(strip.placement = "outside") 

dev.off()

figure_safety2 <- figure_safety %>% 
  mutate(
    Estimator = case_when(Estimator == "KM" ~ "KM",
                          Estimator == "AJ: 1 Competing Event" ~ "AJ: 1",
                          Estimator == "AJ: 2 Competing Events" ~ "AJ: 2"),
    Estimator = factor(Estimator, 
                       levels = c("AJ: 2",
                                  "AJ: 1",
                                  "KM"))
  )


pdf("Primary safety outcome - RDs - 1-6.pdf", height = 6, width = 3.1)

ggplot(figure_safety2) +
  geom_point(aes(y = Estimator, x = RD)) +
   geom_errorbarh(aes(y = Estimator, xmin = RD_LCL, xmax = RD_UCL), height = 0.1, color = "black") +
  geom_label(aes(y = Estimator, x = 1, 
                 label = scales::number(rd_absolute_bias, accuracy = 0.1)), 
             fill = "white", label.size = NA, family="Calibri", size=3.7, color="black") +
  geom_vline(data = unique_rd_values, aes(xintercept = sga_truth_rd), 
             linetype = "dashed", 
             color = "red") +
  facet_wrap(scenario ~ ., ncol = 1, strip.position = "left") +
  labs(y = "Scenario", 
       x = "Risk Difference (per 100)",
       fill = "Treatment Group", 
       ) +
  scale_x_continuous(limits = c(-15, 2),
                     breaks = seq(-15, 0, by  = 5)) +
  # scale_y_discrete(labels = rep("", 36)) +
  theme_minimal() +
  theme(legend.position = "bottom",
        axis.text.y = element_text(size = 14, family = "Calibri", color="black"),
        axis.text.x = element_text(size = 14, family = "Calibri", color="black"),
        strip.text = element_text(size = 14, family = "Calibri", color="black"),
        legend.text = element_text(size = 14, family = "Calibri", color="black"),
        axis.title.x = element_text(size=14, family = "Calibri", color="black"),
        axis.title.y = element_text(size=14, family = "Calibri", color="black"),
        plot.title = element_text(size=14, family = "Calibri", color="black"),
        #axis.title.y=element_blank(),
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", linewidth = 0.3),
        panel.grid.minor.x = element_line(color = "grey70", linewidth = 0.3),
        ) +
  theme(strip.placement = "outside") 

dev.off()



pdf("Primary safety outcome - RRs - 1-6.pdf", height = 6, width = 3.1)

ggplot(figure_safety2) +
  geom_point(aes(y = Estimator, x = RR)) +
  geom_errorbarh(aes(y = Estimator, xmin = RR_LCL, xmax = RR_UCL), height = 0.1, color = "black") +
  geom_label(aes(y = Estimator, x = 1.03, label = scales::number(rr_absolute_bias, accuracy = 0.01)), fill = "white", label.size = NA, family="Calibri", size=3.7, color="black") +
  geom_vline(data = unique_rd_values, aes(xintercept = sga_truth_rr), linetype = "dashed", color = "red") +
  facet_wrap(scenario ~ ., ncol = 1, strip.position = "left") +
  labs(y = "Scenario", 
       x = "Risk Ratio (log-scale)",
       fill = "Treatment Group", 
       ) +
  scale_x_continuous(
    trans='log10',
    limits = c(0.7, 1.05),
    breaks = seq(0.7, 1.05, by = 0.1),
    # breaks = scales::trans_breaks("log10", function(x) 10^x),
    labels = scales::label_number(accuracy = 0.01)
  ) +
  theme_minimal() +
  theme(legend.position = "bottom",
        axis.text.y = element_text(size = 14, family = "Calibri", color="black"),
        axis.text.x = element_text(size = 14, family = "Calibri", color="black"),
        strip.text = element_text(size = 14, family = "Calibri", color="black"),
        legend.text = element_text(size = 14, family = "Calibri", color="black"),
        axis.title.x = element_text(size=14, family = "Calibri", color="black"),
        axis.title.y = element_text(size=14, family = "Calibri", color="black"),
        plot.title = element_text(size=14, family = "Calibri", color="black"),
        axis.title.x.bottom = element_text(hjust = 0.4),
        #axis.title.y=element_blank(),
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", linewidth = 0.3),
        panel.grid.minor.x = element_line(color = "grey70", linewidth = 0.3),
        strip.placement = "outside"
        )  

dev.off()

```

## Scenarios 7-12

```{r}


z <- qnorm(0.975)


## Now, we only want 1 set of KM estimates for each scenario and then each of teh AJ estimates. 
## Split this up

km_safety <- figure_data %>% 
  filter(Outcome != "Composite") %>%  
  # filter(scenario <= 3) %>% 
  select(scenario, contains("km")) %>% 
  unique() %>% 
  mutate(Estimator = "KM") %>% 
  rename(risk_non_initiators = km_r0,
         risk_non_initiators_ucl = km_r0_ucl,
         risk_non_initiators_lcl = km_r0_lcl,
         risk_initiators = km_r1,
         risk_initiators_ucl = km_r1_ucl,
         risk_initiators_lcl = km_r1_lcl,
         RD = km_rd,
         RR = km_rr,
         RD_LCL = km_rd_lcl,
         RD_UCL = km_rd_ucl,
         RR_LCL = km_rr_lcl,
         RR_UCL = km_rr_ucl)

aj_safety <- figure_data %>% 
  filter(Outcome != "Composite") %>% 
  # filter(scenario <= 3) %>% 
  select(scenario, Outcome, contains("aj")) %>% 
  mutate(Estimator = ifelse(grepl("1", Outcome),
                            "AJ: 1 Competing Event",
                            "AJ: 2 Competing Events")) %>% 
  select(-Outcome) %>% 
  rename(risk_non_initiators = aj_r0,
         risk_non_initiators_ucl = aj_r0_ucl,
         risk_non_initiators_lcl = aj_r0_lcl,
         risk_initiators = aj_r1,
         risk_initiators_ucl = aj_r1_ucl,
         risk_initiators_lcl = aj_r1_ucl,
         RD = aj_rd,
         RR = aj_rr,
         RD_LCL = aj_rd_lcl,
         RD_UCL = aj_rd_ucl,
         RR_LCL = aj_rr_lcl,
         RR_UCL = aj_rr_ucl)

figure_safety <- bind_rows(km_safety, aj_safety) %>% 
  left_join(truth, by = "scenario") %>% 
  filter(scenario > 6) %>% 
  mutate(
    composite_truth_r1 = 100*composite_truth_r1,
    composite_truth_r0 = 100*composite_truth_r0,
    composite_truth_rd = 100*composite_truth_rd,
    sga_truth_r1 = 100*sga_truth_r1,
    sga_truth_r0 = 100*sga_truth_r0,
    sga_truth_rd = 100*sga_truth_rd
  ) %>% 
  mutate(rd_absolute_bias = round(abs(sga_truth_rd - RD),1),
         rr_absolute_bias = round(abs(log(sga_truth_rr) - log(RR)), 2)) %>% 
  mutate(Estimator = factor(Estimator,
                            levels = c("AJ: 2 Competing Events",
                                       "AJ: 1 Competing Event",
                                       "KM")))

# Create a data frame with unique composite_truth_rd values for each facet
unique_rd_values <- figure_safety %>% 
  group_by(scenario) %>% 
  summarise(sga_truth_rd = unique(sga_truth_rd),
            sga_truth_r0 = unique(sga_truth_r0),
            sga_truth_r1 = unique(sga_truth_r1),
            sga_truth_rr = unique(sga_truth_rr))




#####
## Print out the risks

pdf("Primary safety outcome - risks - 7-12.pdf", height = 6, width = 4.25)

ggplot(figure_safety) +
  # Incidence bars for Drug A and Comparator
  geom_bar(aes(y = Estimator, x = risk_initiators, fill = "Initiators"), stat = "identity", 
           position = position_nudge(y = 0.2), width = 0.4) +
  geom_bar(aes(y = Estimator, x = risk_non_initiators, fill = "Non-Initiators"), stat = "identity", 
           position = position_nudge(y = -0.2), width = 0.4) +
  # Confidence Intervals for Initiators
  geom_errorbar(aes(y = Estimator, xmin = risk_initiators_lcl, xmax = risk_initiators_ucl), 
                width = 0.2, position = position_nudge(y = 0.2)) +
  # Confidence Intervals for Non-Initiators
  geom_errorbar(aes(y = Estimator, xmin = risk_non_initiators_lcl, xmax = risk_non_initiators_ucl), 
                width = 0.2, position = position_nudge(y = -0.2)) +
  facet_wrap(scenario ~ ., ncol = 1, strip.position = "left") +
  geom_point(data = figure_safety, aes(y = Estimator, x = sga_truth_r1),
             position = position_nudge(y = 0.2), size = 2, shape = 23, fill = "white", color = "black") +
  geom_point(data = figure_safety, aes(y = Estimator, x = sga_truth_r0),
             position = position_nudge(y = -0.2), size = 2, shape = 23, fill = "white", color = "black") +
  labs(y = "Scenario", 
       x = "Absolute Risk (%)",
       fill = "Treatment Group", 
       ) +
  theme_minimal() +
  theme(legend.position = "bottom",
        # axis.title.x.top = element_text(hjust = 0.97),
        axis.text.y = element_text(size = 14, family = "Calibri", color="black"),
        axis.text.x = element_text(size = 14, family = "Calibri", color="black"),
        strip.text = element_text(size = 14, family = "Calibri", color="black"),
        legend.text = element_text(size = 14, family = "Calibri", color="black"),
        axis.title.x = element_text(size=14, family = "Calibri", color="black"),
        axis.title.y = element_text(size=14, family = "Calibri", color="black"),
        plot.title = element_text(size=14, family = "Calibri", color="black"),
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", linewidth = 0.3),
        panel.grid.minor.x = element_line(color = "grey70", linewidth = 0.3),
        ) +
  scale_fill_manual(values = c("Initiators" = "#E69F00", "Non-Initiators" = "#0072B2")) +
  theme(strip.placement = "outside") 

dev.off()

figure_safety2 <- figure_safety %>% 
  mutate(
    Estimator = case_when(Estimator == "KM" ~ "KM",
                          Estimator == "AJ: 1 Competing Event" ~ "AJ: 1",
                          Estimator == "AJ: 2 Competing Events" ~ "AJ: 2"),
    Estimator = factor(Estimator, 
                       levels = c("AJ: 2",
                                  "AJ: 1",
                                  "KM"))
  )


pdf("Primary safety outcome - RDs - 7-12.pdf", height = 6, width = 3.1)

ggplot(figure_safety2) +
  geom_point(aes(y = Estimator, x = RD)) +
   geom_errorbarh(aes(y = Estimator, xmin = RD_LCL, xmax = RD_UCL), height = 0.1, color = "black") +
  geom_label(aes(y = Estimator, x = 4.5, 
                 label = scales::number(rd_absolute_bias, accuracy = 0.1)), 
             fill = "white", label.size = NA, family="Calibri", size=3.7, color="black") +
  geom_vline(data = unique_rd_values, aes(xintercept = sga_truth_rd), 
             linetype = "dashed", 
             color = "red") +
  facet_wrap(scenario ~ ., ncol = 1, strip.position = "left") +
  labs(y = "Scenario", 
       x = "Risk Difference (per 100)",
       fill = "Treatment Group", 
       ) +
  scale_x_continuous(limits = c(-7.5, 5),
                     breaks = seq(-7.5, 2.5, by  = 5)) +
  # scale_y_discrete(labels = rep("", 36)) +
  theme_minimal() +
  theme(legend.position = "bottom",
        axis.text.y = element_text(size = 14, family = "Calibri", color="black"),
        axis.text.x = element_text(size = 14, family = "Calibri", color="black"),
        strip.text = element_text(size = 14, family = "Calibri", color="black"),
        legend.text = element_text(size = 14, family = "Calibri", color="black"),
        axis.title.x = element_text(size=14, family = "Calibri", color="black"),
        axis.title.y = element_text(size=14, family = "Calibri", color="black"),
        plot.title = element_text(size=14, family = "Calibri", color="black"),
        axis.title.x.bottom = element_text(hjust = 0.4, color="black"),
        # axis.title.y=element_blank(),
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", linewidth = 0.3),
        panel.grid.minor.x = element_line(color = "grey70", linewidth = 0.3),
        ) +
  theme(strip.placement = "outside") 

dev.off()



pdf("Primary safety outcome - RRs - 7-12.pdf", height = 6, width = 3.1)

ggplot(figure_safety2) +
  geom_point(aes(y = Estimator, x = RR)) +
  geom_errorbarh(aes(y = Estimator, xmin = RR_LCL, xmax = RR_UCL), height = 0.1, color = "black") +
  geom_label(aes(y = Estimator, x = 1.08, label = scales::number(rr_absolute_bias, accuracy = 0.01)), fill = "white", label.size = NA, family="Calibri", size=3.7, color="black") +
  geom_vline(data = unique_rd_values, aes(xintercept = sga_truth_rr), linetype = "dashed", color = "red") +
  facet_wrap(scenario ~ ., ncol = 1, strip.position = "left") +
  labs(y = "Scenario", 
       x = "Risk Ratio (log-scale)",
       fill = "Treatment Group", 
       ) +
  scale_x_continuous(
    trans='log10',
    limits = c(0.85, 1.1),
    breaks = seq(0.85, 1.05, by = 0.1),
    # breaks = scales::trans_breaks("log10", function(x) 10^x),
    labels = scales::label_number(accuracy = 0.01)
  ) +
  theme_minimal() +
  theme(legend.position = "bottom",
        axis.text.y = element_text(size = 14, family = "Calibri", color="black"),
        axis.text.x = element_text(size = 14, family = "Calibri", color="black"),
        strip.text = element_text(size = 14, family = "Calibri", color="black"),
        legend.text = element_text(size = 14, family = "Calibri", color="black"),
        axis.title.x = element_text(size=14, family = "Calibri", color="black"),
        axis.title.y = element_text(size=14, family = "Calibri", color="black"),
        plot.title = element_text(size=14, family = "Calibri", color="black"),
        axis.title.x.bottom = element_text(hjust = 0.4, color="black"),
        # axis.title.y=element_blank(),
        panel.grid.major.y = element_blank(),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", linewidth = 0.3),
        panel.grid.minor.x = element_line(color = "grey70", linewidth = 0.3),
        ) +
  theme(strip.placement = "outside") 

dev.off()

```

# Calculate Stats for Manuscript

## Primary Efficacy Outcome

```{r}


## Efficacy Outcome
composite_outcome %>% 
  select(-c(contains("_se"))) %>% 
  filter(scenario <= 6) %>% 
  mutate(
    km_r0_bias = km_r0 - truth_r0,
    km_r1_bias = km_r1 - truth_r1
  ) %>% 
  ungroup() %>% 
  summarize(
    avg_km0_bias = mean(km_r0_bias),
    avg_km1_bias = mean(km_r1_bias),
    avg_km_rd_bias = mean(abs(km_rd_bias)),
    avg_km_rr_bias = mean(abs(km_lnrr_bias))
  )


  



```

## Primary Safety Outcome - Only Fetal Death as Competing Event

```{r}

sga_fd_only %>% 
  select(-c(contains("_se"))) %>% 
  mutate(
    km_r0_bias = km_r0 - truth_r0,
    km_r1_bias = km_r1 - truth_r1,
    aj_r0_bias = aj_r0 - truth_r0,
    aj_r1_bias = aj_r1 - truth_r1
  ) %>% 
  ungroup() %>% 
  summarize(
    avg_km0_bias = mean(km_r0_bias),
    avg_km1_bias = mean(km_r1_bias),
    avg_km_rd_bias = mean(abs(km_rd_bias)),
    avg_km_rr_bias = mean(abs(km_lnrr_bias)),
    avg_aj0_bias = mean(aj_r0_bias),
    avg_aj1_bias = mean(aj_r1_bias),
    avg_aj_rd_bias = mean(abs(aj_rd_bias)),
    avg_aj_rr_bias = mean(abs(aj_lnrr_bias))
  ) %>% 
  mutate(
    avg_km_bias = (avg_km0_bias + avg_km1_bias) / 2,
    avg_aj_bias = (avg_aj0_bias + avg_aj1_bias) / 2
  )

```

## Primary Safety Outcome - All Competing Events

```{r}

safety_all <- sga_all %>% 
  select(-c(contains("_se")))

```

## Median Gestational Age at Outcome

```{r}

combined_describe_outcomes %>% 
  filter(strata == "Total") %>% 
  group_by(preg_outcome_final) %>% 
  summarize(
    median_ga = median(median)
  )


```
